requirements.txt
astroid==2.4.1
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
google-api-core==1.19.1
google-auth==1.16.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.2
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
